Migration des CPT_IMP vers des FLOAT (au lieu des INT)
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@920 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ Include/Reseaux_option_compteur_imp.h
@@ -37,7 +37,7 @@
 
  struct CMD_TYPE_OPTION_COMPTEUR_IMP
   { guint  id_mnemo;                                                        /* Id unique du mnemo associé */
-    guint  valeur;                                                                  /* Valeur du compteur */
+    gdouble valeur;                                                                 /* Valeur du compteur */
     guint  unite;                                                                         /* Km, h, ° ... */
     guint  type;                                                                          /* Km, h, ° ... */
                                                                                         /* Vient du mnemo */
